backend/pyproject.toml
[project]
name = "personamimic-backend"
version = "0.1.0"
description = "PersonaMimic AI Orchestrator Backend"
authors = [{ name = "PersonaMimic Team" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.1",
    "faiss-cpu>=1.8.0",
    "sentence-transformers>=2.5.1",
    "cryptography>=42.0.5",
    "pydantic>=2.6.3",
    "aiofiles>=23.2.1",
    "python-multipart>=0.0.9",
    "ollama>=0.1.7",
    "googlesearch-python>=1.2.3",
    "requests>=2.31.0",
    "sqlalchemy>=2.0.28",
    "psutil>=5.9.8",
    "beautifulsoup4>=4.12.3",
    "scikit-learn>=1.4.1",
    "pandas>=2.2.1",
    "speechbrain>=0.5.16",
    "torch>=2.2.1",
    "torchaudio>=2.2.1",
    "soundfile>=0.12.1",
    "chromadb>=0.4.24",
    "openai-whisper>=20231117",
    "sentry-sdk[fastapi]>=2.0.0",
    "logfire[fastapi]>=0.40.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",
    "edge-tts>=6.1.0",
    "mcp>=1.27.0",
    "transformer-lens>=2.17.0",
    "sae-lens>=6.39.0",
    "bitsandbytes>=0.49.2",
    "safetensors>=0.7.0",
    "angr>=9.2.209",
    "langgraph>=0.0.30",
    "langsmith>=0.1.0",
    "langchain>=0.1.0",
    "nats-py>=2.7.0",
    "semgrep>=1.65.0",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM", "ASYNC"]
ignore = ["UP006", "UP035", "B008", "E501"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.10",
]
